moved vhost and mount tests
